-Created UtilityFunctions.kt to store repeated functions which are used across multiple fragments. -Pulled out the logic from FavoritesFragment.kt, WishlistFragment.kt, and ExploreViewModel.kt, which is used to prep the data for the recycler view, and put it in UtilityFunctions.kt
